+++ docs/release-notes.md
@@ -7,3 +7,4 @@ A non-exhaustive list of what has changed in a more readable form than a commit 
  - Implemented Persistence of Beta-Blockers after Heart Attack Measure
  - Implemented All-cause Readmissions (PCR) Measure
  - Implemented Childhood Immunization Measure
+ - Implemented Annual Monitoring for Patients on Diuretics Measure